% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kodom_components.R
\name{collect_boot_phi}
\alias{collect_boot_phi}
\title{Sign-align bootstrap eigenfunctions to the main fit}
\usage{
collect_boot_phi(boot_fits, ref_fit, slot, K, n_args)
}
\description{
Sign-align bootstrap eigenfunctions to the main fit
}
\keyword{internal}
