% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kodom_fit.R
\name{fpca_long}
\alias{fpca_long}
\title{Reshape an N x T FPCA reconstruction matrix to long (id, time, value)}
\usage{
fpca_long(mat, argvals, ids, id, time, value)
}
\description{
Reshape an N x T FPCA reconstruction matrix to long (id, time, value)
}
\keyword{internal}
