% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_kodom_branch.R
\docType{data}
\name{GeomKodomBranch}
\alias{GeomKodomBranch}
\title{Geom for geom_kodom_branch}
\format{
A ggproto object.
}
\usage{
GeomKodomBranch
}
\description{
Extends \link[ggplot2:Geom]{ggplot2::GeomPath} to draw branching trajectory plots. Drops the
\code{id} and \code{medication} columns in \code{setup_data()} so ggplot2 does not warn
about unknown aesthetics at draw time.
}
\keyword{internal}
