% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_kodom_heatmap.R
\docType{data}
\name{GeomKodomHeatmap}
\alias{GeomKodomHeatmap}
\title{Geom for geom_kodom_heatmap}
\format{
A ggproto object.
}
\usage{
GeomKodomHeatmap
}
\description{
Extends \link[ggplot2:Geom]{ggplot2::GeomTile} with kodom defaults. Drops \code{id} in
\code{setup_data()} to avoid an unknown-aesthetic warning from ggplot2.
}
\keyword{internal}
