% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_kodom_periodic.R
\docType{data}
\name{GeomKodomPeriodic}
\alias{GeomKodomPeriodic}
\title{Geom for geom_kodom_periodic}
\format{
A ggproto object.
}
\usage{
GeomKodomPeriodic
}
\description{
Extends \link[ggplot2:Geom]{ggplot2::GeomPath} to draw concentric "star trail" arcs.
Interpolates colours/alpha in the (angle, radius) space that the stat
already produces, then delegates the polar projection to
\code{coord_polar(theta = "x")}. This means standard ggplot2 scales and
coordinates work as expected: use \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}} for
time-axis labels, \code{\link[ggplot2:coord_radial]{ggplot2::coord_polar()}} for direction, and
\code{\link[=theme_kodom_periodic]{theme_kodom_periodic()}} for a clean panel.
}
\keyword{internal}
