% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_helpers.R
\name{make_pc_label}
\alias{make_pc_label}
\title{Build a single PC facet label}
\usage{
make_pc_label(k, pve, show_pve = TRUE)
}
\arguments{
\item{k}{PC index.  Use 0 for the mean panel.}

\item{pve}{Numeric vector of proportions of variance explained.}

\item{show_pve}{Annotate with PVE? (default \code{TRUE}).}
}
\description{
Build a single PC facet label
}
\keyword{internal}
