% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_kodom_periodic.R
\docType{data}
\name{StatKodomPeriodic}
\alias{StatKodomPeriodic}
\title{Stat for geom_kodom_periodic}
\format{
A ggproto object.
}
\usage{
StatKodomPeriodic
}
\description{
Assigns lane positions and converts lane rank to an Archimedean spiral
radius (\code{y = inner_radius + lane * lane_width + spiral_fraction * total_time / period}).
Time \code{x} is left continuous, and relies on \code{coord_polar(theta = "x")} to natively wrap
the data in a 360 degree space for values extending past the \code{period}.
}
\keyword{internal}
